Moved sessions observer to separate class
