+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/mobile/MobileActiveController.kt
@@ -23,10 +23,6 @@ class MobileActiveController(
 ): SessionsController(mRootActivity, mViewMvc, mSessionsViewModel, mLifecycleOwner, mSettings),
     SessionsViewMvc.Listener {
 
-    init {
-        mSessionsLiveData = loadSessions()
-    }
-
     override fun loadSessions(): LiveData<List<SessionWithStreamsDBObject>> {
         return mSessionsViewModel.loadMobileActiveSessionsWithMeasurements()
     }
